% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alienAggregate.R
\name{alienAggregate}
\alias{alienAggregate}
\title{Aggregate alien object.}
\usage{
alienAggregate(object, aggregateCol, keepCols = NULL, idSuffix = "id_")
}
\arguments{
\item{object}{An object of class \code{alienData}.}

\item{aggregateCol}{A character string or integer identifying the column of \code{dfNodes} to be used to aggregate interactions.}

\item{keepCols}{Vector of strings or integer identifying the columns to be
kept along with \code{aggregateCol}.}

\item{idSuffix}{Suffix of the new identifiers.}
}
\value{
An object of the class \code{alienPredict} with the desired new level of interactions.
}
\description{
Aggregate the interactions of an \code{alienData} object
according to a a given aggregateCol of \code{dfNodes}.
}
\examples{
data(plantsPol)
plantsPolSp <- alienAggregate(plantsPol, 'idSp', 'type')
# plantsPol2 <- plantsPol
# plantsPol2$dfNodes$var1 <- runif(nrow(plantsPol2$dfNodes))
# plantsPolSp <- alienAggregate(plantsPol2, 'idSp', 'var1')
}
\author{
Kevin Cazelles
}
\keyword{manip}
