% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alienPredict.R
\name{alienPredict}
\alias{alienPredict}
\title{Format alien output.}
\usage{
alienPredict(logLik, netEstim, ...)
}
\arguments{
\item{logLik}{Likelihood log-transformed.}

\item{netEstim}{a probabilistic network.}

\item{...}{other arguments to be appended to the \code{alienPredict} object.}
}
\value{
An object of the class \code{alienPredict} is returned .
}
\description{
Format alien output and return an object of class \code{alienPredict}.
}
\details{
\code{alienPredict} aims at structuring the output of alien package to ease
cross-comparison among methods.
}
\author{
Kevin Cazelles & Steve Vissault
}
\keyword{classes}
\keyword{manip}
