% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.alienData.R
\name{as.alienData}
\alias{as.alienData}
\title{Format data for the \code{alien} class}
\usage{
as.alienData(dfSpecies, dfInteract, trait = NULL, phylo = NULL,
  taxo = NULL, siteEnv = NULL, traitSp = NULL, traitInd = NULL,
  phy = NULL, dfSite = NULL, dfOcc = NULL, verbose = TRUE)
}
\arguments{
\item{dfSpecies}{A data frame with a least two colomuns. A given row describes one
observation. The site indentifiers (\code{idSp}, must be unique) has to come first
and the species identifier (\code{idSp}) in second. Two
extra columns could be provided by the user: \code{idTime} adding a timestamp to the observations and
\code{idInd}, an identifier of individus. If these two extra columns
are not supplied, \code{NA} will be added. Columns must be correctly ordered.}

\item{dfInteract}{A data frame with three columns which contains interaction at the finest level
(individus or species). The first two columns are \code{idFrom} and \code{idTo} and
determine the sens of the interaction. idFrom and \code{idTo} are unique identifier
of species or individu documented in the \code{idObs} data frame. Finaly, the thrid
column is the strength of the interaction (Please see details).}

\item{trait}{A square symmetric matrix of 0s and 1s that define co-occurence patterns among pairs of species. If this matrix is not provided, the co-occurence matrix is derived from the coAbund matrix else the \code{idObs} dataframe (see return section).}

\item{phylo}{A square symmetric matrix that includes any types of values, defining co-abundance patterns among pairs of species. TODO: Not implemented yet.}

\item{taxo}{A matrix or a data frame where each column is a descriptor of the sites. TODO: siteEnv should cover the possibility that environmental variables could be taken at several times - link to idTime in idObs?.}

\item{dfSite}{Logical}

\item{dfOcc}{Logical}

\item{verbose}{Logical}

\item{binary}{Logical}

\item{directed}{Logical}
}
\value{
An object of the class \code{alienData} is returned by \code{as.alienData}.
TODO: Declare more accurately the structure of object returned by the function. What has been generated from the function.
}
\description{
This function formats the data and returns an object of class alienData.
}
\details{
\code{idObs} is used to check consistency and prevent errors among unique
identifiers of each alienData arguments.

The strength of the interactions defined in the third column of
\code{interactPair} can be a 0 if no direct interaction has been observed
(defined as true absence of interaction) or any numerical value.
WARNING: interactSp is aggregating the information at the species level using the sum function by default.
Undocumented interactions among species or individus will be assumed as NA
by default.
}
\author{
Guillaume Blanchet, Kevin Cazelles & Steve Vissault
}
\keyword{classes}
\keyword{manip}
