% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bartomeus}
\alias{bartomeus}
\title{Plant and pollinator example network}
\format{A data frame with X variables}
\usage{
data(bartomeus)
}
\description{
A dataset containing plant and pollinator lengths ...
}
\details{
\itemize{
  \item plant: Plant latin name
  \item nectar_holder_depth_mm: Plant nectar holder depth in mm
  \item pollinator: Pollinator latin name
  \item IT_mm: Pollinator Intertegular span in mm
  \item family: Pollinator family
  \item cover: Plant cover from independent transects
  }
}
\author{
Ignasi Bartomeus
}
\keyword{datasets}
