% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_belmaker_et_at_2009.R
\name{bayesian_example}
\alias{bayesian_example}
\title{Fit an example Bayesian model}
\usage{
bayesian_example(
  data,
  time_col,
  alien_col,
  native_col,
  native_pool_size,
  priors,
  ...
)
}
\arguments{
\item{data}{a data.frame containing columns for alien and native discoveries}

\item{time_col}{the name of time since invasion column in the data}

\item{alien_col}{the name of alien discovery column in the data}

\item{native_col}{the name of native discovery column in the data}

\item{native_pool_size}{an integer of the assumed number of native species}

\item{priors}{a named numeric vector containing the elements:  \tabular{ll}{
\code{b0_mu} \tab prior for the b0 mean.\cr
\code{b1_mu} \tab prior for the b1 mean.\cr
\code{b0_sd} \tab prior for the b0 sd.\cr
\code{b1_sd} \tab prior for the b1 sd.\cr
}}

\item{...}{additional arguments passed to cmdstanr's \code{sample}.}
}
\value{
the fit value
}
\description{
Fit an example Bayesian heirarchical model to data, using the proportion of
undiscovered alien species to the total number of undiscovered species (natives and aliens)
as the probability of detecting a new alien species. The model is described in full in Belmaker et al (2009),
with some modifications described in Buba et al ().
}
\details{
This function is mainly used to demonstrate how a Bayesian model can be implemented when aiming
to estimate the underlying introduction rates.
It can be used for user data if data on both alien and native discoveries are present and
,model assumptions are met.
}
\examples{
data(medfish)
example_priors <- c(b0_mu = 0, b1_mu = 0, b0_sd = 0.01, b1_sd = 0.001)
bayesian_example(data = medfish,
   time_col = "time",
   aliens_col = "aliens",
   native_col = "natives",
   native_pool_size = 600,
   priors = example_priors,
   chains = 3, parallel_chains = 3)
}
\seealso{
\code{\link[=snc]{snc()}} in which other external data can be used as dependent variables for either \eqn{\mu_t} or \eqn{\Pi_{st}}.
}
