% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estNichePars.R
\name{estNichPars}
\alias{estNichPars}
\title{Fit a model using Genaralized Simulated Annealing (SA) functions.}
\usage{
estNichPars(model, Tlevel1, Tlevel2, mean_Tlevel1, sd_Tlevel1, pars = c(a0 =
  0, a1 = 0, b0 = 0, b1 = 0), par_lo = c(a0 = -5000, a1 = -5000, b0 = -5000,
  b1 = -5000), par_hi = c(a0 = 5000, a1 = 5000, b0 = 5000, b1 = 5000),
  max.time = 1800)
}
\arguments{
\item{model}{One of the three models implemented in niche.prob: integrated.model, niche.model or neutral.model.}

\item{Tlevel1}{Vector of trait values of the first interaction partner.}

\item{Tlevel2}{Vector of trait values of the second interaction partner.}

\item{mean_Tlevel1}{Mean of trait values of the first interaction partner. Can be weighted or not,
and can use independent information on the trait distribution to be calculated.}

\item{sd_Tlevel1}{Standard deviation of trait values of the first interaction partner. Can be weighted or not,
and can use independent information on the trait distribution to be calculated.}

\item{pars}{a vector of the form c(a0 = 0, a1 = 0, b0 = 0, b1 = 0) with the initial parameters.}

\item{max.time}{value in seconds to stop the estimation of params. Set to 1800 (30 minuts)}

\item{pars_lo}{a vector of the form c(a0 = 0, a1 = 0, b0 = 0, b1 = 0) with the lower limits that the
parameters can reach. You can use a priosy information to constrain the posible values of the parameters
(e.g. the slope, a1, has to be positive)}

\item{pars_hi}{a vector of the form c(a0 = 0, a1 = 0, b0 = 0, b1 = 0) with the higher limits that the
parameters can reach. You can use a priosy information to constrain the posible values of the parameters}
}
\value{
a vector with the estimated parameters of the form c(a0 = x, a1 = y, b0 = z, b1 = w)
}
\description{
Fit the paramaters of a given model to data using SA. Can be slow for large datasets.
}
\note{
More information on this method at https://github.com/ibartomeus/trait_match
}
\author{
Dominique Gravel and Ignasi Bartomeus
}

