% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.bayesreg.R
\name{extract.bayesreg}
\alias{extract.bayesreg}
\title{Extract posterior distributions from Bayesian Regression Model}
\usage{
extract.bayesreg(x)
}
\arguments{
\item{x}{A JAGS model file from fit.bayesreg()}
}
\description{
A helper function for getting the chains from a JAGS model object outputted from fit.bayesreg
}
\author{
Ben Weinstein
}
\references{
Bartomeus et al. 2016. Functional Ecology.
}

