% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.bayesreg.R
\name{fit.bayesreg}
\alias{bayesian}
\alias{fit.bayesreg}
\alias{regression}
\title{Hierarchical Bayesian Regression for estimating trait-matching in species interaction data}
\usage{
fit.bayesreg(dat, algorithm = "Binomial")
}
\arguments{
\item{dat}{A data frame with columns named 'I','J','Interactions','TraitI','TraitJ'.
I: The identity of the species on the upper level (eg. pollinators)
J: The identity of the species on the lower level (eg. plants)
Interactions: 1 (Observed) or 0 (non-detected)}

\item{algorithm}{A string argument, either 'Intercept' or 'Poisson', see details
Must have JAGS v 4.0 > to run. Install jags [here](http://mcmc-jags.sourceforge.net/)
#' @details
Intercept Model
For each pair of species i interaction with species j
$$ Obs_{i,j} \sim Binom(\rho_{i,j})$$
$$ logit(\rho_{i,j}) = \alpha_{i,j} $$

With a hierarchical relationship among species i (eg. pollinators)
$$\alpha_{i,j} \sim Normal(\alpha_\mu,\alpha_\sigma)$$
#' Intercept Model
For each pair of species i interaction with species j
$$ Obs_{i,j} \sim Binom(\rho_{i,j})$$
$$ logit(\rho_{i,j}) = \alpha_{i,j} + \beta_{i,j}$$

With a hierarchical relationship among species intercepts and slopes i (eg. pollinators)
$$\alpha_{i,j} \sim Normal(\alpha_\mu,\alpha_\sigma$$
$$\beta_{i,j} \sim Normal(\beta_\mu,\beta_\sigma$$}
}
\value{
A jags model obect (see package R2Jags)
}
\author{
Ben Weinstein
}

