% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBayesreg.R
\name{fitBayesreg}
\alias{bayesian}
\alias{fitBayesreg}
\alias{regression}
\title{Hierarchical Bayesian Regression for estimating trait-matching in species interaction data}
\usage{
fitBayesreg(dat, algorithm = "Binomial", draws = 10000)
}
\arguments{
\item{dat}{A data frame with columns named 'I','J','Interactions','TraitI','TraitJ'.
I: The identity of the species on the upper level (eg. pollinators)
J: The identity of the species on the lower level (eg. plants)
Interactions: 1 (Observed) or 0 (non-detected)}

\item{algorithm}{A string argument, either 'Intercept' or 'Poisson', see details
Must have JAGS v 4.0 > to run. Install jags [here](http://mcmc-jags.sourceforge.net/)}

\item{draws}{The number of MCMC draws to run before taking 100 samples from the posterior distribution}
}
\value{
A jags model obect (see package R2Jags)
}
\details{
Intercept Model
For each pair of species i interaction with species j
\eqn{ Obs_i,j ~ Binom(\rho_i,j)}
\eqn{ logit(\rho_{i,j}) = \alpha_i,j}

With a hierarchical relationship among species i (eg. pollinators)
\deqn{\alpha_{i,j} ~ Normal(\alpha_\mu,\alpha_\sigma)}
#' Intercept Model
For each pair of species i interaction with species j
\deqn{ Obs_{i,j} ~ Binom(\rho_{i,j})}
\deqn{ logit(\rho_{i,j}) = \alpha_{i,j} + \beta_{i,j}}

With a hierarchical relationship among species intercepts and slopes i (eg. pollinators)
\deqn{\alpha_{i,j} ~ Normal(\alpha_\mu,\alpha_\sigma}
\deqn{\beta_{i,j} ~ Normal(\beta_\mu,\beta_\sigma}
}
\author{
Ben Weinstein
}

