% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDirectCentrality.R
\name{fitDirectCentrality}
\alias{fitDirectCentrality}
\title{Direct matching centrality}
\usage{
fitDirectCentrality(data, algorithm, ...)
}
\arguments{
\item{data}{a dataframe with a first column named I (the interaction) and extra columns with covariates (can be named at will)}

\item{algorithm}{one of: logistic, poisson, RF}

\item{...}{other params passed to algorithm}
}
\description{
Currently deals with logistic and poisson regression, and random forest
}
\author{
Dominique Gravel
}
\references{

}

