% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDMC.R
\name{fitDMC}
\alias{fitDMC}
\title{Fit direct matching centrality}
\usage{
fitDMC(formula, data, class = NULL, family = NULL, traits = NULL,
  step = FALSE, ...)
}
\arguments{
\item{formula}{formula to define the model.}

\item{data}{an object of the class \code{\link{alienData}}}

\item{class}{Method to use to estimate the model. Either "randomForest" (\link[randomForest]{randomForest}) or "glm" (\code{\link[stats]{glm}}).}

\item{family}{The family of the response variable. See \link[stats]{family}, or the choices available.}

\item{\dots}{Other parameters passed to either \link[randomForest]{randomForest} or \link[stats]{glm}.}
}
\description{
Fit direct matching centrality model
}
\author{
Dominique Gravel, Steve Vissault, F. Guillaume Blanchet
}
