% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGLM.R
\name{fitGLM}
\alias{fitGLM}
\title{Fit direct matching centrality using generalized linear model}
\usage{
fitGLM(data, formula, family = NULL, spRandom = FALSE, ...)
}
\arguments{
\item{data}{an object of the class \code{\link{alienData}}}

\item{formula}{A one-sided formula specifying how the different traits from both sets of species should be used to estimate species interactions. Default is to include all terms additively, with quadratics for quantitative terms, and  interactions between all pairs of traits.}

\item{family}{The family of the response variable. See \link[stats]{family}, or the choices available.}

\item{spRandom}{Logical. Whether species are used as a random effect. Default is FALSE.}

\item{\dots}{Other parameters passed to \link[stats]{glm}.}
}
\value{
An object with a class alienFit and a class fitGLM.
}
\description{
Fit direct matching centrality model using generalized linear model
}
\details{
The generalized linear model is designed to be used on bipartite network where traits are available for both sets of species interacting in the network. It should not be used otherwise.

This function unfold the adjacency matrix and uses it as the response variable. As explanatory variables, the traits for each sets of species are repeated to match the length of the unfolded adjacency matrix but also the position.

If there are NAs in the adjacency matrix, the function will omit these values in the estimation of the model.

Also, the default \code{formula} option results in many explanatory variables being constructed, which may result in overparamerisation. In this respect, the default option is complete but often too much for the data at hand.
}
\author{
F. Guillaume Blanchet, Dominique Gravel, Steve Vissault
}
