% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitIMC.R
\name{fitIMC}
\alias{fitIMC}
\title{Fit indirect matching centrality model}
\usage{
fitIMC(data, d = 1, verbose = TRUE, control = list())
}
\arguments{
\item{data}{An object of the class \code{\link{alienData}}.}

\item{d}{Numeric. The dimension to the latent traits.}

\item{verbose}{Logical. Whether information on the progress of the analysis is reported in the console.}

\item{control}{List passed to \code{\link[GenSA]{GenSA}} to control the behavior of the algorithm.}
}
\value{
An object with a class alienFit and a class fitIMC.
}
\description{
This method estimate latent traits to model interactions in an adjacency matrix.
}
\details{
WRITE MORE AND BETTER STUFF
WRITE MORE AND BETTER STUFF
WRITE MORE AND BETTER STUFF

This method uses latent traits to model interactions in an adjacency matrix. The first step requires the definition of two sets of interacting
species: set1 and set2 (respectively of size \code{nset1} and \code{nset2}).
The first category of latent trait are the matching traits. (to be developped)
The second category include the centrality terms. (to be developped).
Latents traits are built under specific topological constraints:
\itemize{
\item 1- all vectors belongs to the orthogonal basis of the unit vector.
\item 2- all matching vector of a particular set of species are orthogonal to each other.
}
}
\references{
Rohr, R. P. & Bascompte, J. (2014) Components of Phylogenetic Signal in Antagonistic and Mutualistic Networks. Am. Nat. 184, 556--564.

Rohr, R. P., Naisbit, R. E., Mazza, C. & Bersier, L.-F. (2016) Matching-centrality decomposition and the forecasting of new links in networks. Proc. R. Soc. B Biol. Sci. 283, 20152702.
}
\author{
Kevin Cazelles, Dominique Gravel and F. Guillaume Blanchet
}
