% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitKNN.R
\name{fitKNN}
\alias{fitKNN}
\title{Fit using K-nearest neighbour}
\usage{
fitKNN(
  data,
  distFrom = "jaccard",
  distTo = "jaccard",
  distTraitFrom = NULL,
  distTraitTo = NULL,
  weight = 0.5,
  nNeig,
  phylo = FALSE
)
}
\arguments{
\item{data}{An object of the class alienData, see \code{\link{alienData}}.}

\item{distFrom}{Character string defining which distance (or dissimilarity) to apply on the "From" species. Check \code{\link[vegan]{vegdist}} for the distances to choose from. Default is "jaccard".}

\item{distTo}{Character string defining which distance (or dissimilarity) to apply on the "To" species. Check \code{\link[vegan]{vegdist}} for the distances to choose from. Default is "jaccard".}

\item{distTraitFrom}{Character string defining which distance (or dissimilarity) to apply on the traits of the "From" species. Check \code{\link[vegan]{vegdist}} for the distances to choose from. If the value is left to NULL, \code{traitDistFrom} (or \code{phyloDistFrom}) in the data (\code{\link{alienData}}) is used directly. Default is NULL.}

\item{distTraitTo}{Character string defining which distance (or dissimilarity) to apply on the "To" species. Check \code{\link[vegan]{vegdist}} for the distances to choose from. If the value is left to NULL, \code{traitDistTo} (or \code{phyloDistTo}) in the data (\code{\link{alienData}}) is used directly. Default is NULL.}

\item{weight}{Numeric. Defines the contribution of the traits (or phylogeny) in the analysis. Must be between 0 and 1. Default is 0.5.}

\item{nNeig}{Integer  defining how many neighbours to consider.}

\item{phylo}{Logical. Whether phylogenetic information should be used instead of traits to measure neighbourhood. Default is FALSE.}
}
\value{
An object with a class alienFit and a class fit4corner.
}
\description{
Model adegency matrix using K-nearest neighbour approach
}
\details{
This function should only be used for bipartite adjency matrice.

The function is designed in such a way that if the argument \code{distTraitFrom} is defined it will build a distance matrix using \code{traitFrom} in the \code{\link{alienData}} object even if \code{distTraitFrom} is available in \code{\link{alienData}} object. The same is true for the argument \code{distTraitTo}.

The argument \code{weight} defines the important of traits (or phylogeny) in the analysis. If a weight of 0 is given, the traits (or phylogeny) are assumed to have no importance. Conversely, if \code{weight} is 1 the traits (or phylogeny) are given their full importance in the analysis.

If \code{phylo} is TRUE, the cophenetic distance is used to calculate the distance between pairs of species.

When ranking the species to find the \code{nNeig} nearest neighbour, in case of ties the argument \code{tie.method} in the \code{\link{rank}} function is set to "first". Also, when ranking the species, it is assumed that the species is not a neighbour of itself.

NAs were removed in the calculation of the distances whenever they were present, but also in the calculation of the interaction probability. For species where all distance values are NAs, the returned interactions probability will be 0.
}
\author{
F. Guillaume Blanchet and Dominique Gravel
}
