% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitNicheBin.R
\name{fitNicheBin}
\alias{fitNicheBin}
\title{This function use a pair of traits.}
\usage{
fitNicheBin(data, thresh)
}
\arguments{
\item{data}{a dataset with two columns named Tj and Ti}

\item{thresh}{quantiles to be estimated (tau in qr() function)}
}
\description{
By convention, i is the target layer and j is the opposed layer
Data: a data frame with Ti as traits of the layer i and Tj traits of the layer j
each row correspond to a pair of species with known interaction
}
\author{
Dominique Gravel
}
\references{

}

