% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitRF.R
\name{fitRF}
\alias{fitRF}
\title{Fit direct matching centrality using Random Forest}
\usage{
fitRF(data, formula, ...)
}
\arguments{
\item{data}{an object of the class \code{\link{alienData}}}

\item{formula}{A one-sided formula specifying how the different traits from both sets of species should be used to estimate species interactions.}

\item{\dots}{Other parameters passed to \link[randomForest]{randomForest}.}
}
\value{
An object with a class alienFit and a class fitRF.
}
\description{
Fit direct matching centrality model using Random Forest
}
\details{
The Random Forest model is designed to be used on bipartite network where traits are available for both sets of species interacting in the network. It should not be used otherwise.

This function unfold the adjacency matrix and uses it as the response variable. As explanatory variables, the traits for each sets of species are repeated to match the length of the unfolded adjacency matrix but also the position.

If there are NAs in the adjacency matrix, the function will omit these values in the estimation of the model.
}
\author{
F. Guillaume Blanchet, Dominique Gravel, Steve Vissault
}
