% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdjacencyMatrix.R
\name{getAdjacencyMatrix}
\alias{getAdjacencyMatrix}
\title{Compute the adjacency matrix for a given alienData object}
\usage{
getAdjacencyMatrix(object, bipartite = FALSE, binary = FALSE, threshold = 0)
}
\arguments{
\item{object}{An object of class \code{alienData}.}

\item{bipartite}{Logical. Whether the adjacency matrix should be constructed for a associated with a
bipartite network? Default is \code{FALSE}.}

\item{binary}{Logical. Should the interactions be binary (see details)?
Otherwise the \code{value} column of \code{edge} is used. Default is \code{FALSE}.}

\item{threshold}{A positive-real number used to determine binary interactions (ignored if \code{binary} is \code{FALSE}).}
}
\value{
An adjacency matrix of class \code{matrix}.
}
\description{
Computes the adjacency matrix based on
the \code{edge} par of an \code{alienData} object.
}
\details{
By default \code{getAdjacencyMatrix} creates a square matrix including all
nodes found in the \code{node} data frame of \code{object}. Then, it reads
\code{edge} to fill the matrix out. If bipartite is \code{TRUE} then only the
nodes found in the \code{from} column of \code{edge} are used to name
the rows of the adjacency matrix. Similarly, the \code{to} nodes become the
columns names.

Currently if there are several values for the same interaction
\code{getAdjacencyMatrix} sums them unless \code{binary} is \code{TRUE}.
}
\author{
Kevin Cazelles, F. Guillaume Blanchet
}
\keyword{adjacency}
\keyword{matrix}
