% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConnectance.R
\name{getConnectance}
\alias{getConnectance}
\title{Compute the connectance of a probabilistic network.}
\usage{
getConnectance(netEstim)
}
\arguments{
\item{netEstim}{a probabilistic network.}
}
\value{
An object of the class \code{alienPredict} is returned .
}
\description{
Compute the connectance of a probabilistic network and its variance.
}
\examples{
getConnectance(rep(.5,10))
}
\references{
Poisot, T. et al. The structure of probabilistic networks. Methods Ecol. Evol. 7, 303–312 (2016).
}
\author{
Kevin Cazelles & Steve Vissault
}
\keyword{connectance}
