% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEdgesList.R
\name{getEdgesList}
\alias{getEdgesList}
\title{Extract the edges list from an adjacency matrix.}
\usage{
getEdgesList(x, bipartite = TRUE, ignore = 0)
}
\arguments{
\item{x}{an adjacency matrix.}

\item{bipartite}{A logical. Is the adjacency matrix associated with a
bipartite network? Default is set to \code{TRUE}.}

\item{ignore}{a vector whose values will be regarded as an absence of interaction.}
}
\value{
A data frame of three columns \code{idFrom}, \code{idTo} and \code{value}.
}
\description{
\code{getEdgesList} extracts the edges list from an adjacency matrix.
}
\details{
Row and column names of the matrix are used to identify the nodes of the edges
list. If at least one of these attributes is \code{NULL} then names are
assigned as follows: if \code{bipartite} is \code{TRUE} two different sets
of names are used, otherwise the same set of names is used. In the last
scenario, last names (in alphabetical order) are droped from the smaller set.
Note that names of matrix columns are treated as \code{idFrom} nodes and
names of matrix rows are treated as \code{idTo} nodes.
}
\author{
Kevin Cazelles
}
\keyword{adjacency}
\keyword{edges}
\keyword{list}
\keyword{matrix}
