% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTraitMatrix.R
\name{getTraitMatrix}
\alias{getTraitMatrix}
\title{Compute the trait matrix (or matrices) for a given alienData object}
\usage{
getTraitMatrix(object, level = "species")
}
\arguments{
\item{object}{An object of class \code{alienData}.}

\item{level}{Either 'individual' or 'species'. The ecological level on which the traits should be extracted. Default is 'species'.}
}
\description{
Computes the trait matrix (or matrices) using \code{dfNodes} and \code{nmTrait} of an \link{alienData} object.
}
\author{
F. Guillaume Blanchet
}
