% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWideDF.R
\name{getWideDF}
\alias{getWideDF}
\title{Create a dataframe of the wide format.}
\usage{
getWideDF(data, nodes, sites)
}
\arguments{
\item{data}{an object of the class alienData or an adjacency matrix.}

\item{nodes}{a logical. Should nodes' information be included.}

\item{sites}{a logical. Should sites' information be included.}
}
\value{
An object of class \code{alienPredict}.
}
\description{
Create a dataframe of the wide format.
}
\examples{
getWideDF()
}
\author{
Kevin Cazelles
}
