% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iEat.r
\name{iEat_bin}
\alias{iEat_bin}
\title{Instance-based machine learning method to predict biotic interactions}
\usage{
iEat_bin(S0, S1, S2 = S1, sourceSim, targetSim = sourceSim, K = 5,
  minSim = 0.3, minWt = 1, predict = "full algorithm")
}
\arguments{
\item{S0}{Matrix, catalogue of empirical data used to infer predictions}

\item{S1}{Vector of taxa forming networking for which topology is predicted}

\item{S2}{Vector of taxa in S1 for which we wish to predict resources (if unspecified, S2 == S1 and the whole network is predicted)}

\item{sourceSim}{Matrix (numeric), source similarity matrix between S1 taxa and the union of S0 and S1 taxa, structure sourceSim[unique(S0,S1), S1]}

\item{targetSim}{Matrix (numeric), source similarity matrix between S1 taxa and the union of S0 and S1 taxa, structure sourceSim[unique(S0,S1), S1] (if unspecified, targetSim == sourceSim and no similarity distinction between resources and consumers)}

\item{K}{Integer, how many neighbours for K nearest neighbour evaluation}

\item{minSim}{Integer, minimum similarity value accepted to consider taxa as similar (implemented to avoid unrealistic interactions)}

\item{minWt}{Integer, minimum weight for candidate source to become a predicted source}

\item{predict}{String, specifies whether the predictions are made from the "full algorithm", the "catalogue" or the "predictive" contribution. If unspecified, predict == 'full algorithm'. See Beauchesne et al. (2016) for more details. If predict == 'catalogue', the methodology corresponds to the approach presented by Gray et al. (2015).}
}
\value{
A dataframe with source taxa for which target predictions are made, target infered from catalogue data (empirical) and target infered from KNN algorithm
}
\author{
David Beauchesne
}

