% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCPredict.R
\name{IMCPredict}
\alias{IMCPredict}
\title{Format IMC output.}
\usage{
IMCPredict(logLik, netEstim, ...)
}
\arguments{
\item{logLik}{Likelihood log-transformed.}

\item{netEstim}{a probabilistic network.}

\item{...}{other arguments to be appended to the \code{alienPredict} object.}
}
\value{
An object of the class \code{IMCPredict} is returned.
}
\description{
Format IMC output and return an object of class \code{IMCPredict}.
}
\details{
\code{IMCPredict} aims at structuring the output of the IMC function.
}
\author{
Kevin Cazelles & Steve Vissault
}
\keyword{internal}
