% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kopelke}
\alias{kopelke}
\title{Salix / Gallers / Parasitoids network}
\format{An alienData object containing the data decribed in Kopelke et al. (2017)}
\usage{
data(kopelke)
}
\description{
This is an alienData object built from the data set described in Kopelke et al.
(2017), i.e. a single community type sampled across Europe: willow species
(Salix), willow-galling sawflies (Hymenoptera: Tenthredinidae: Nematinae:
Euurina), and their natural enemies (hymenopteran parasitoids and coleopteran,
lepidopteran, dipteran, and hymenopteran inquilines).
}
\references{
Kopelke, J.-P. et al. Food-web structure of willow-galling sawflies and
their natural enemies across Europe. Ecology 38, 42–49 (2017).
}
\author{
Ignasi Bartomeus
}
\keyword{datasets}
