% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{medfish}
\alias{medfish}
\title{Discoveries Of Native And Alien Fish Species In The Eastern Mediterranean Sea}
\format{
A data frame with 60 rows and 4 columns:
\describe{
\item{year}{Year of record}
\item{time}{Time since the first record in the records data}
\item{natives}{Number of newly recorded native species in corresponding year}
\item{aliens}{Number of newly recorded alien species in corresponding year}
}
}
\source{
\url{https://esajournals.onlinelibrary.wiley.com/doi/full/10.1890/07-1904.1}
}
\usage{
medfish
}
\description{
The data used in Belmaker et al (2009) describing discoveries of native and alien species in the Mediterranean Sea
between the years 1927–2017.
}
\references{
Belmaker, J., Brokovich, E., China, V., Golani, D., & Kiflawi, M. (2009). Estimating the rate of biological introductions: Lessepsian fishes in the Mediterranean. Ecology, 90(4), 1134–1141. https://doi.org/10.1890/07-1904.1
}
\keyword{datasets}
