% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsToJags.R
\name{binomialToJags}
\alias{binomialToJags}
\alias{interceptToJags}
\alias{multinomialToJags}
\alias{poissonToJags}
\title{Write models as JAGS file.}
\usage{
binomialToJags(filename)

interceptToJags(filename)

multinomialToJags(filename)

poissonToJags(filename)
}
\arguments{
\item{filename}{A connection, or a character string naming the file to print to.}
}
\value{
A \code{.jags} file including the desired model.
}
\description{
Write one of the available model into a \code{.jags} file.
}
\section{Functions}{
\itemize{
\item \code{binomialToJags}: Binomial model.

\item \code{interceptToJags}: Intercept model.

\item \code{multinomialToJags}: Multinomial model.

\item \code{poissonToJags}: Poisson model.
}}
\author{
Ben Weistein
}

