% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkBayesReg.R
\name{networkBayesReg}
\alias{networkBayesReg}
\title{Generate network statistics from a hierarchical bayesian model of species interactions}
\usage{
networkBayesReg(x, indices)
}
\arguments{
\item{x}{A predicted model object from predict.bayesreg}

\item{indices}{a vector of network statistics corresponding to network statistics (see details)}
}
\value{
A data frame with the desired metrics for each of the draws of the network.
}
\description{
Generate network statistics from a hierarchical bayesian model of species interactions
}
\details{
Network statistics are calculated using the fucntion networklevel from the package bipartite, see ?networklevel for a list of available options.
}

