% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEM.build.alienData.R
\name{PEM.build.alienData}
\alias{PEM.build.alienData}
\title{Construct Phylogenetic Eigenvector Maps with \code{\link{alienData}}}
\usage{
\method{PEM.build}{alienData}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{alienData}}. See details.}

\item{\dots}{Arguments passed to \code{\link[MPSEM]{PEM.build}}. See details.}
}
\value{
An object of class \code{\link{alienData}}.
}
\description{
This function is essentially a wrapper around \code{\link[MPSEM]{PEM.build}}. It is designed to use the phylogenetic information in an \code{\link{alienData}} object to construct phylogenetic eigenvector maps and replace (or add) them as traits in a new \code{\link{alienData}} object.
}
\details{
For this function it is essential that the \code{\link{alienData}} object includes phylogenetic information, otherwise the function will send an error message. However, it is not necessary for the \code{\link{alienData}} object to include traits.

The arguments passed to \code{\link[MPSEM]{PEM.build}} are typically that ones used for tunning when constructing the phylogenetic eigenvector maps.
}
\references{
Guénard, G., Legendre, P., and Peres-Neto, P. 2013. Phylogenetic eigenvector maps (PEM): a framework to model and predict species traits. \emph{Methods in Ecology and Evolution} \strong{4} 1120-1131.
}
