% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPred.R
\name{plotPred}
\alias{plotPred}
\title{Plot the predicted model output with data points on top.}
\usage{
plotPred(pars, Tlevel1, Tlevel2, xlab = "Trait level 2",
  ylab = "Trait level 1", legend = TRUE, pch = 19, ...)
}
\arguments{
\item{pars}{Parameters of the model obtained with fit.niche.prob function (alpha0, alpha 1, beta 0 and beta 1)}

\item{Tlevel1}{Vector of trait values of the first interaction partner.}

\item{Tlevel2}{Vector of trait values of the second interaction partner.}

\item{xlab}{label for x axes}

\item{ylab}{label for y axes}

\item{legend}{Logical value indicating if the legend should be plotted. Default = TRUE}

\item{pch}{Default = 19. Use '.' when lots of data available.}

\item{...}{other parameters to be passed to plot.}
}
\value{
A plot
}
\description{
This is just a pre-built function for easy visualization. Other options are possible.
Only works with the default gaussian model (integrated and niche model).
}
\note{
See more here: https://github.com/ibartomeus/trait_match
}
\author{
Dominique Gravel and Ignasi Bartomeus
}

