% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predDirectCentrality.R
\name{predDirectCentrality}
\alias{predDirectCentrality}
\title{Predict direct matching centrality}
\usage{
predDirectCentrality(newdata, model, replicates)
}
\arguments{
\item{newdata}{a new dataframe with a first column named I (the interaction) and extra columns with covariates (same name as used to fit)}

\item{model}{the output of fit.direct.centrality}

\item{replicates}{Number of interactions to predict based on the interaction probability values}
}
\description{
It generates predicted interactions based on new data.
}
\author{
Dominique Gravel
}
\references{

}

