% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vazquez2009.R
\name{predict.vazquez2009}
\alias{predict.vazquez2009}
\title{Simulate a network from a given probability matrix following null model proposed in Vazquez 2009}
\usage{
\method{predict}{vazquez2009}(web, n, keep.species = TRUE, rep.cell = T)
}
\arguments{
\item{web}{A I * J matrix of probabilities (binary) or total counts (quantitative)}

\item{n}{Total number of observations}

\item{keep.species}{}

\item{rep.cell}{}
}
\value{
A predicted matrix of interactions
}
\description{
Simulate a network from a given probability matrix following null model proposed in Vazquez 2009
}
\details{
This is simply a wrapper function for the mgen function in the package bipartite, included here as part of the quantitative tests. See ?bipartite:mgen for details.
}

