% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictDMC.R
\name{predictDMC}
\alias{predictDMC}
\title{prediction based on direct matching centrality method}
\usage{
predictDMC(model, data = NULL, type = "response")
}
\arguments{
\item{model}{a model applied on a alienData object.}

\item{type}{kinf od response desired.}

\item{newdata}{the data used for predictions.}
}
\description{
predict species interaction based on the direct matching centrality model
}
\author{
Dominique Gravel & Steve Vissault
}
