% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predNicheBin.R
\name{predNicheBin}
\alias{predNicheBin}
\title{This function use a pair of traits.}
\usage{
predNicheBin(models, newdata)
}
\arguments{
\item{models}{The output of fit.niche.bin}

\item{newdata}{a dataframe with with Ti as traits of the layer i and Tj traits of the layer j
each row correspond to a pair of trait for which we want to know if there is an
interaction happenning}
}
\description{
By convention, i is the target layer and j is the opposed layer
Data: a data frame with Ti as traits of the layer i and Tj traits of the layer j
each row correspond to a pair of species with known interaction
}
\author{
Dominique Gravel
}
\references{

}

