% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predNicheProb.R
\name{predNicheProb}
\alias{predNicheProb}
\title{Predict new data points using traitmatch function generated with fit.niche.prob}
\usage{
predNicheProb(pars, Tlevel1, Tlevel2, replicates = 100)
}
\arguments{
\item{pars}{Parameters of the model obtained with fit.niche.prob function (alpha0, alpha 1, beta 0 and beta 1)}

\item{Tlevel1}{Vector of trait values of the first interaction partner.}

\item{Tlevel2}{Vector of trait values of the second interaction partner.}

\item{replicates}{Number of interactions to predict based on the interaction probability values}
}
\value{
A list. First item is the probability for each predicted interactions.
The next elements are as many predicted interaction as replicates.
}
\description{
It generates predicted interactions based on new data.
}
\note{
See more here: https://github.com/ibartomeus/trait_match
}
\author{
Dominique Gravel and Ignasi Bartomeus
}

