% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.niche.prob.R
\name{residuals.niche.prob}
\alias{residuals.niche.prob}
\title{get the residuals from a fitted niche.prob}
\usage{
\method{residuals}{niche.prob}(pars, Tlevel1, Tlevel2)
}
\arguments{
\item{pars}{Parameters of the model obtained with fit.niche.prob function (alpha0, alpha 1, beta 0 and beta 1)}

\item{Tlevel1}{Original vector of trait values of the first interaction partner.}

\item{Tlevel2}{Original vector of trait values of the second interaction partner.}
}
\value{
a vector with the residual values for each interaction.
}
\description{
get the residuals from a fitted niche.prob
}
\author{
Ignasi Bartomeus and Dominique Gravel
}

