% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residualsNicheProb.R
\name{residualsNicheProb}
\alias{residualsNicheProb}
\title{get the residuals from a fitted nicheProb}
\usage{
residualsNicheProb(pars, Tlevel1, Tlevel2)
}
\arguments{
\item{pars}{Parameters of the model obtained with fitNicheProb function (alpha0, alpha 1, beta 0 and beta 1)}

\item{Tlevel1}{Original vector of trait values of the first interaction partner.}

\item{Tlevel2}{Original vector of trait values of the second interaction partner.}
}
\value{
a vector with the residual values for each interaction.
}
\description{
get the residuals from a fitted nicheProb
}
\author{
Ignasi Bartomeus and Dominique Gravel
}

