% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsquare.R
\name{rsquare}
\alias{rsquare}
\title{Pseudo-\eqn{R^2}{R2}}
\usage{
rsquare(alienFit)
}
\arguments{
\item{alienFit}{An object of class \code{alienFit}.}
}
\description{
Calculates Efron's pseudo-\eqn{R^2}{R2} for \code{alienFit} objects
}
\details{
Efron's pseudo-\eqn{R^2}{R2} is calculated as:

\deqn{R^2 = 1 - \frac{(y_i-\hat{\pi})^2}{(y_i-\bar{y})^2}{R2=1-((y-pihat)^2)/((y-ybar)^2)}}.

where \eqn{\hat{\pi}}{pihat} is the model predicted values calculated on the scale of the response variable.

This version of pseudo-\eqn{R^2}{R2} was chosen over the others because it can be calculated for all types of models implemented in this package.
}
\references{
Efron, B. (1978) Regression and ANOVA with Zero-One Data: Measures of Residual Variation., \emph{Journal of the American Statistical Association} \strong{73}, 113-121.
}
