% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{salixGal}
\alias{salixGal}
\title{Salix - Galler - parasitoids data}
\format{
an object of class alienData
}
\usage{
data(salixGal)

data(galPara)
}
\description{
The Salix - galler - parasitoids data was originally published by Kopelke et al. (2017) for which additional information on phylogeny was added. The full data was used to perform the analysis in Wooton et al. (2021).
}
\references{

}
\keyword{datasets}
