% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sfestuary}
\alias{sfestuary}
\title{Discoveries Of Introduced Species In The San Francisco Estuary (California, USA)}
\format{
An object of class \code{integer} of length 145.
}
\source{
\url{https://esajournals.onlinelibrary.wiley.com/doi/abs/10.1890/03-3102}

\url{https://repository.library.noaa.gov/view/noaa/40918}
}
\usage{
sfestuary
}
\description{
The data provided by Solow and Costello (2004) describing discoveries of introduced species in the San Francisco estuary (California, USA)
between the years 1850–1995. These data were compiled by Cohen and Carlton (1995); see also Cohen and Carlton (1998). \cr
\cr
The original data in PDF format is described in full in Cohen and Carlton (1995) and can be viewed here: https://repository.library.noaa.gov/view/noaa/40918
}
\references{
Cohen, A. N., and J. T. Carlton. 1995. Nonindigenous aquatic species in a United States estuary: a case study of the biological invasions of the San Francisco Bay and Delta. U.S. Fish and Wildlife Service, Washington, D.C., USA. https://repository.library.noaa.gov/view/noaa/40918

Cohen, A. N., and J. T. Carlton. 1998. Accelerating invasion rate in a highly invaded estuary. Science 279: 555– 558. https://www.science.org/doi/10.1126/science.279.5350.555

Solow, A. R., and Costello, C. J. 2004. Estimating the rate of species introductions from the discovery record. Ecology, 85(7), 1822–1825. https://esajournals.onlinelibrary.wiley.com/doi/abs/10.1890/03-3102
}
\keyword{datasets}
