% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.bayesreg.R, R/sim.traitmatch.R
\name{extract.bayesreg}
\alias{extract.bayesreg}
\alias{extract_par}
\alias{getpars}
\alias{sim.traitmatch}
\alias{sim.traitmatch.R}
\alias{trait-matching}
\title{Extract posterior distributions from Bayesian Regression Model}
\usage{
extract_par(x)

sim.traitmatch(size_x, size_y, traits_x, traits_y, beta_sigma = 0.4,
  alpha_sigma = 0)
}
\arguments{
\item{x}{2bedone}

\item{size_x}{The number of species in the lower level (eg. plants)}

\item{size_y}{The number of species in the upper level (eg. pollinators)}

\item{traits_x}{Vector of trait values for species in the lower level.}

\item{traits_y}{Vector of trait values for species in the lower level.}

\item{beta_sigma}{Variance among species in trait-matching slope.}

\item{alpha_sigma}{Variance among species in trait-matching intercept.}
}
\description{
A helper function for getting the chains from a JAGS model object outputted from fit.bayesreg

A helper functional for generating the probability of interaction among hypothetical species based on their traits.
The trait-matching function is currently |Trait_i - Trait_j|
}
\author{
Ben Weinstein

Ben Weinstein
}
\references{
Bartomeus et al. 2016. Functional Ecology.

Bartomeus et al. 2016. Functional Ecology.
}

