% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPhyloTraits.R
\name{simPhyloTraits}
\alias{simPhyloTraits}
\title{Traits generator based on evolutionary models.}
\usage{
simPhyloTraits(nsps, ntraits = 1, model = c("lambda", "delta", "OU"), param)
}
\arguments{
\item{nsps}{integer giving the number of species to be returned.}

\item{ntraits}{integer giving the number of traits to be simulated.}

\item{model}{a characer string giving the evolutionary model used to simulated traits.}

\item{param}{parameter values of the model to be used (see \link[geiger]{rescale}).}
}
\value{
A species-by-traits matrix.
}
\description{
This function simulates traits values according different evolutionary models.
}
\examples{
mattr <- simPhyloTraits(100, 3, 'lambda', 0)
}
\author{
Ignacio Morales Castilla
Kevin Cazelles
}

