% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTraitMatch.R
\name{simTraitMatch}
\alias{simTraitMatch}
\alias{trait-matching}
\title{Simulating trait-matching among species}
\usage{
simTraitMatch(size_x, size_y, traits_x, traits_y, beta_sigma = 0.1,
  alpha_sigma = 0, replicates = 1, type = "binary")
}
\arguments{
\item{size_x}{The number of species in the lower level (eg. plants)}

\item{size_y}{The number of species in the upper level (eg. pollinators)}

\item{traits_x}{Vector of trait values for species in the lower level.}

\item{traits_y}{Vector of trait values for species in the lower level.}

\item{beta_sigma}{Variance among species in trait-matching slope.}

\item{alpha_sigma}{Variance among species in trait-matching intercept.}

\item{replicates}{The number of times the network was sampled. For example, Replicate = 1 will yield I*J data points}

\item{type}{string binary or quantitative network?}
}
\description{
A helper functional for generating the probability of interaction among hypothetical species based on their traits.
The trait-matching function is currently |Trait_i - Trait_j|
}
\author{
Ben Weinstein
}
\references{
Bartomeus et al. 2016. Functional Ecology.
}

