% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sourceMC}
\alias{sourceMC}
\title{sourceMC}
\arguments{
\item{nbsp1}{Number of species belonging to the first set.}

\item{nbsp2}{Number of species belonging to the second set.}
}
\value{
Returns the eigen values associated to the vector and eigen values
transition matrix of the markov chain to be solved.
}
\description{
Linear algebra operations needed for fitMC.
}
\author{
Kevin Cazelles
}
