% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.alienData.R
\name{summary.alienData}
\alias{print.summary.alienData}
\alias{summary.alienData}
\title{Format data for the \code{alien} class}
\usage{
\method{summary}{alienData}(object, ...)

\method{print}{summary.alienData}(x, ...)
}
\arguments{
\item{object}{An object of class \code{alienData}.}

\item{...}{Further arguments to be passed from or to other methods.}

\item{x}{An object of class \code{summaryAlienData}.}
}
\description{
These functions are all \code{methods} for class \code{alienData}
or \code{summary.alienData}.
}
\author{
Kevin Cazelles
}

