% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snc.R
\name{summary_snc}
\alias{summary_snc}
\title{Summarize a Solow and Costello Model Fit}
\usage{
summary_snc(object)
}
\arguments{
\item{object}{an object of class "snc", usually, a result of a call to \code{snc}}
}
\value{
A data.frame containing the model estimates, standard error, and the probability of the true value being 0
under the given estimates and errors.
}
\description{
Summarize a Solow and Costello Model Fit
}
\examples{
\donttest{
data(sfestuary)
example_model <- snc(sfestuary)
summary(example_model)
}
}
