% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjur.R
\name{tjur}
\alias{tjur}
\title{Calculate Tjur's D}
\usage{
tjur(object)
}
\arguments{
\item{object}{An object of class \code{alienfit}.}
}
\value{
A numerical value presenting a Tjur's D.
}
\description{
Calculate Tjur's D for an object of class \code{alienfit}.
}
\details{
This function is designed for presence-absence data and will send an error when other data type are used.
}
\author{
F. Guillaume Blanchet and Kate Wootton
}
\keyword{univar}
