% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{traitsBasedJointCooc}
\alias{traitsBasedJointCooc}
\title{Generate traits-based joint co-occurrence.}
\arguments{
\item{env}{a numerical vector including the environmental values.}

\item{metaweb}{a numerical matrix depicting the biotic interactions.}

\item{alpha}{a real positive number determining the importance of interactions: the
higher it is the more interactions impact occurrence values.}

\item{trait1}{a numerical vector giving the environmental optima of species.}

\item{trait2}{a optional numerical vector giving the probabilities of presence at the optimal value}

\item{trait3}{a optional numerical vector giving the generalities of species.}
}
\value{
A numeric gradient x species matrix of presence probabilities of the
interacting species associated to the environmental values.
}
\description{
This function generates occurrence probabilities of a set of interacting
species given for a given set of environnmental values.
}
\author{
Kevin Cazelles
}

