% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{webFromTraitMatching}
\alias{webFromTraitMatching}
\title{Generate a web using a trait matching approach.}
\arguments{
\item{nbsp}{an integer giving the number of species considered.}

\item{trait1}{an numeric vector giving the first trait to be matched with the second one. If NULL (the default values) then values are drawn from an uniform distribution.}

\item{trait2}{an numeric vector giving the second trait to be matched with the first one. if NULL (the default values) then values are drawn from an uniform distribution.}

\item{trait3}{an numeric vector giving the first trait to be matched if NULL (the default values) then values are drawn from an uniform distribution.}

\item{trait4}{an numeric vector (between 0 and 1) giving the first trait to be matched if NULL (the default values) then values are drawn from an uniform distribution.}
}
\value{
A logical matrix describing pairwise interactions. A given line
describe the diet of a given species while a column decribes the set of
predator accociated to a particular species.
}
\description{
This function generates a food webs based on 1- a trait matching between
two traits, 2- a generality trait and 3- an efficiency trait.
}
\details{
For any couple of species (i,j) the probability that i interact with
(\emph{e.g.} i feeds upon j) is as follows:
\deqn{p(M_{i,j}=1) = trait4_i e^{-\left(-\frac{trait1_i-trait2_j}{trait3_i}\right)^2}}{%
      p(M_{i,j}=1) = trait4[i] exp(-((trait1[i]-trait2[j])/trait3[i])^2)}
The length of all trait vectors must equal the number of species otherwise
an error is returned.
This apporach has been used with feeding traits in a version revisited of
niche model implemented (see \lik[pkg:alienR]{webFromNicheModel}) by Williams
\emph{et al.} (2010) where \code{trait4} is common to all species.
}
\author{
Kevin Cazelles
}
\references{
Williams, R. J., Anandanadesan, A. & Purves, D. (2010) The probabilistic
niche model reveals the niche structure and role of body size in a complex
food web. PLoS One 5.
}

