% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightFcts.R
\name{weightFcts}
\alias{weightFcts}
\alias{weightedMean}
\title{weighted mean and sd}
\usage{
weightedMean(x, w)
}
\arguments{
\item{x}{vector to compute the mean over.}

\item{w}{a vector of same length with the weigths of each element.}
}
\value{
a value with the weighted mean or sd.
}
\description{
just weighted mean and sd.
}
\examples{
weightedMean(c(2,3,4), c(1,1,2))
weightedSd(c(2,3,4), c(2,1,2))

}

