% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alienData.R
\name{alienData}
\alias{alienData}
\title{Formatting data and return an \code{alienData} object}
\usage{
alienData(
  adjMat,
  traitFrom = NULL,
  traitTo = NULL,
  phyloFrom = NULL,
  phyloTo = NULL,
  traitDistFrom = NULL,
  traitDistTo = NULL,
  phyloDistFrom = NULL,
  phyloDistTo = NULL
)
}
\arguments{
\item{adjMat}{An adjancency matrix. The rows (From) species are influencing the column (To) species.}

\item{traitFrom}{A data.frame containing the traits of the row (From) species.}

\item{traitTo}{A data.frame containing the traits of the column (To) species.}

\item{phyloFrom}{An object of class \code{\link[ape]{phylo}} for the row (From) species.}

\item{phyloTo}{An object of class \code{\link[ape]{phylo}} for the column (To) species.}

\item{traitDistFrom}{A dist object containing the distance between pairs of traits of the row (From) species.}

\item{traitDistTo}{A dist object containing the distance between pairs of traits of the column (To) species.}

\item{phyloDistFrom}{A dist object containing phylogenetic distance between pairs of row (From) species.}

\item{phyloDistTo}{A dist object containing phylogenetic distance between pairs of column (To) species.}
}
\value{
An object of class \code{alienData} is returned.
}
\description{
\code{alienData} is used to check the data, if correct
it returns an object of class \code{alienData}.
}
\details{
This function is essentially designed to make sure the names of all components match. The output of this function is at the basis of all the analyses implemented in this package.
}
\author{
F. Guillaume Blanchet, Kevin Cazelles & Steve Vissault
}
\keyword{classes}
\keyword{manip}
