% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.alienData.R
\name{as.alienData}
\alias{as.alienData}
\title{Format data for the \code{alien} class}
\usage{
as.alienData(interactPair = NULL, coOcc = NULL, coAbund = NULL,
  interact = NULL, siteSp = NULL, siteEnv = NULL, traitSp = NULL,
  traitInd = NULL, phylo = NULL, resCon = NULL, location = NULL,
  scaleSiteEnv = TRUE, scaleTrait = TRUE, interceptSiteEnv = TRUE,
  interceptTrait = TRUE)
}
\arguments{
\item{interactPair}{A data.frame with species name in the two first columns (two interacting species), the strength of the interaction in the third column (see details) and location (in space or time) where the species was found in the following columns where species were found (see details).}

\item{coOcc}{A square symmetric matrix of 0s and 1s that define co-occurence patterns among pairs of species.}

\item{coAbund}{A square symmetric matrix that includes any types of values, defining co-abundance patterns among pairs of species.}

\item{interact}{A square non-symmetric matrix that presents the interaction includes any types of values.}

\item{siteSp}{A matrix or a data.frame where each column is a species.}

\item{siteEnv}{A matrix or a data.frame where each column is a descriptor of the sites.}

\item{traitSp}{A matrix or a data.frame where each column is a trait characterizing all species.}

\item{traitInd}{A matric or a data.frame where each column is a trait characterizing an individual.}

\item{phylo}{A square symmetric matrix describing the phylogenetic relationships between pairs of all species (see details).}

\item{resCon}{A matrix or data.frame where rows are resources and columns are consumers.}

\item{location}{A factor defining a sample location (in space or through time) or a data.frame characterizing multiple locations.}

\item{scaleSiteEnv}{Logical. Whether the columns of X should be centred and divided by the standard deviation. Default is TRUE.}

\item{scaleTrait}{Logical. Whether the rows of Tr should be centred and divided by the standard deviation. Default is TRUE.}

\item{interceptSiteEnv}{Logical. Whether a column of 1s should be added to X. Default is TRUE.}

\item{interceptTrait}{Logical. Whether a row of 1s should be added to Tr. Default is TRUE.}
}
\value{
An object of the class \code{alienData} is returned by \code{as.Netdata}.
}
\description{
This functions is used to format the data
}
\details{
The strength of the interactions defined in the third column of \code{interactPair} can be an NA or any numerical value.
In addition, species location (sites), will be converted to a factor. Note that there can be multiple sets of locations that can be considered, each one defined by a different factor.
}
\examples{

}
\author{
F. Guillaume Blanchet
}
\keyword{classes}
\keyword{manip}

