% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.alienData.R
\name{as.alienData}
\alias{as.alienData}
\title{Format data for the \code{alien} class}
\usage{
as.alienData(dfNodes, dfEdges, trait = NULL, phylo = NULL, taxo = NULL,
  siteEnv = NULL, traitSp = NULL, traitInd = NULL, phy = NULL,
  dfSite = NULL, dfOcc = NULL, binary = FALSE, directed = FALSE,
  verbose = TRUE)
}
\arguments{
\item{dfNodes}{A vector or a data frame with at least one column named \code{idNodes} providing
unique identifiers for each species identified within the dataset. The remainig
columns could be either traits or phylogenetic or taxonomic data that must
be specified respectively by \code{trait}, \code{phylo} or \code{taxo}
parameter described below.}

\item{dfEdges}{A data frame with at least two columns: \code{idFrom} and \code{idTo}
descibring the set of edges. If \code{directed} is set to \code{TRUE} is also
describes the direction of the interaction. Two additonnal columns are checked
\code{value} \code{iDsite} which respectively stand for the value associated
with the edge and where the interaction has been observed.}

\item{trait}{A vector indicating columns number (or names) of \code{dfNodes} containing traits data (see \code{Details}).}

\item{phylo}{A vector indicating colums number (or names) of \code{dfNodes} containing phylo data (see \code{Details}).}

\item{taxo}{A vector indicating columns number (or names) of \code{dfNodes} containing taxo data (see \code{Details}).}

\item{dfSite}{A data frame with at least two columns named \code{idFrom}}

\item{dfOcc}{A data frame with at least two columns}

\item{binary}{Logical. If `TRUE`, interactions are turned into a boolean variable (see \code{Details}).}

\item{directed}{Logical. If `TRUE` (default value) the network is considered as directed (see \code{Details}).}

\item{verbose}{Logical. Should extra information be reported on progress?}
}
\value{
An object of the class \code{alienData} is returned by \code{as.alienData}.
TODO: Declare more accurately the structure of object returned by the function. What has been generated from the function.
}
\description{
This function formats the data and returns an object of class alienData.
}
\details{
The user is required to provide specific column names to prevent from
generating errors. Two primary keys \code{idNodes} and \code{idSite} if
argument \code{dfSite} is not null.

trait/phylo/taxo non exlusive a column be three or both. There are associated and help distinguishing methods.

character of factors => as.charcater
The strength of the interactions defined in the third column of

If no values in \code{dfEdges} then values are set to 1 is used.
\code{directed} => methods and => plots
\code{binary} => methods and option
}
\author{
Guillaume Blanchet, Kevin Cazelles & Steve Vissault
}
\keyword{classes}
\keyword{manip}
