% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.alienData.R
\name{as.alienData}
\alias{as.alienData}
\alias{as.alienData.alienData}
\alias{as.alienData.matrix}
\alias{as.alienData.data.frame}
\alias{as.alienData.igraph}
\title{Convert various formats to an alienData object.}
\usage{
as.alienData(x, ...)

\method{as.alienData}{alienData}(x, ...)

\method{as.alienData}{matrix}(x, ...)

\method{as.alienData}{data.frame}(x, ...)

\method{as.alienData}{igraph}(x, ...)
}
\arguments{
\item{x}{Input to as.alienData}

\item{...}{other arguments to be passed to \code{\link[alien]{alienData}}.}
}
\value{
An object of the class \code{alienPredict}.
}
\description{
\code{as.alienData} is a generic function woth method to convert
\code{matrix}, \code{data.frame} and \code{igraph} object into a
an \code{alienData} object.
}
\details{
For objects of class \code{matrix}, see \code{\link[alien]{getEdgesList}}
to learn how the edges list is retrieved from the adjacency matrix.
The same principles are applied to objects of class \code{igraph} except
they are handled using funtions from the \code{igraph} package.
}
\author{
Kevin Cazelles
}
\keyword{classes}
\keyword{method}
