% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.alienData.R
\name{as.alienData}
\alias{as.alienData}
\title{Format data for the \code{alien} class}
\usage{
as.alienData(idObs, interactPair = NULL, coOcc = NULL, coAbund = NULL,
  siteEnv = NULL, traitSp = NULL, traitInd = NULL, phy = NULL,
  scaleSiteEnv = FALSE, scaleTrait = FALSE, interceptSiteEnv = FALSE,
  interceptTrait = FALSE, verbose = TRUE)
}
\arguments{
\item{idObs}{A data frame of a least two colomuns. A given row describes one
observation. The site indentifier (\code{idSite}, must be unique) is provided in the first
and the species identifier (\code{idSp}) in the second. There are two
optional columns: \code{idTime} to add a time stamp to the observations and
\code{idInd} to add a identifier to individuals. Note that if these columns
are emoty, \code{NA} will be added. Columns must be correctly ordered.}

\item{interactPair}{A data frame which contains interaction at the finest level
(individus or species). The first two columns are idFrom and idTo and
determine the sens of the interaction. idFrom and idTo are unique identifier
of species or individu documented in the idObs data frame. Finaly, the thrid
column is the strength of the interaction (Please see details).}

\item{coOcc}{A square symmetric matrix of 0s and 1s that define co-occurence patterns among pairs of species. If this matrix is not provided, the co-occurence matrix is derived from the coAbund matrix else the idObs dataframe (see return section).}

\item{coAbund}{A square symmetric matrix that includes any types of values, defining co-abundance patterns among pairs of species. TODO: Not implemented yet.}

\item{siteEnv}{A matrix or a data frame where each column is a descriptor of the sites. TODO: siteEnv should cover the possibility that environmental variables could be taken at several times - link to idTime in idObs?.}

\item{traitSp}{A matrix or a data frame where each column is a trait characterizing all species. The first column is a unique identifier of the species documented in idObs data.frame.}

\item{traitInd}{A matrix or a data frame where each column is a trait characterizing an individual. The first column is a unique identifier of the individu documented in idObs data.frame.}

\item{phy}{An object of class 'phylo' describing the phylogenetic relationships across species (see details).}

\item{scaleSiteEnv}{Logical. Whether the columns of X should be centred and divided by the standard deviation. Default is TRUE.}

\item{scaleTrait}{Logical. Whether the rows of Tr should be centred and divided by the standard deviation. Default is TRUE.}

\item{interceptSiteEnv}{Logical. Whether a column of 1s should be added to X. Default is TRUE.}

\item{interceptTrait}{Logical. Whether a row of 1s should be added to Tr. Default is TRUE.}

\item{verbose}{Logical. warning and help messages are printed. Default is FALSE.}
}
\value{
An object of the class \code{alienData} is returned by \code{as.alienData}.
TODO: Declare more accurately the structure of object returned by the function. What has been generated from the function.
}
\description{
This function formats the data and returns an object of class alienData.
}
\details{
\code{idObs} is used to check consistency and prevent errors among unique
identifiers of each alienData arguments.

The strength of the interactions defined in the third column of
\code{interactPair} can be a 0 if no direct interaction has been observed
(defined as true absence of interaction) or any numerical value.
Undocumented interactions among species or individus will be assumed as NA
by default.
}
\examples{

}
\author{
F. Guillaume Blanchet & Steve Vissault
}
\keyword{classes}
\keyword{manip}
