% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBayesReg.R
\name{extractBayesreg}
\alias{extractBayesreg}
\title{Extract posterior distributions from Bayesian Regression Model}
\usage{
extractBayesreg(x)
}
\arguments{
\item{x}{A JAGS model file from fit.bayesreg()}
}
\description{
A helper function for getting the chains from a JAGS model object outputted from fit.bayesreg
}
\author{
Ben Weinstein
}
\references{
Bartomeus et al. 2016. Functional Ecology.
}

