% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit4corner.R
\name{fit4corner}
\alias{fit4corner}
\title{Fit fourth corner prediction model}
\usage{
fit4corner(data, formula = NULL, family = NULL, ...)
}
\arguments{
\item{data}{an object of the class \code{\link{alienData}}.}

\item{formula}{A one-sided formula specifying exactly how to model the adjacency matrix as a function of the "From" and "To" traits variables of both sets of species. Default is to include all terms additively, with quadratics for quantitative terms, and all From-by-To traits interactions.}

\item{family}{an object of class \code{\link[stats]{family}} with the caveat that the negative binomial with unknown overdispersion and a log-link can be specified as "negative.binomial", and it is the default..}

\item{\dots}{Other parameters passed to \code{\link[mvabund]{traitglm}}.}
}
\value{
An object with a class alienFit and a class fit4corner.
}
\description{
Fit fourth corner prediction model using the approach proposed by Brown et al. (2014). This function is essentially a wrapper around \code{\link[mvabund]{traitglm}} for it to be used in the context of species interaction data.
}
\details{
Fourth corner models are designed to be used on bipartite network where traits are available for both sets of species interacting in the network. It can not be used otherwise.

Fourth corner models assume that adjacency matrix (species by species matrix) used to perform the analysis has as rows the "From" species and as columns the "To" species.
}
\author{
F. Guillaume Blanchet
}
