% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDirectCentrality.R
\name{fitDirectCentrality}
\alias{fitDirectCentrality}
\title{Fit Direct matching centrality among species or individus}
\usage{
fitDirectCentrality(data, class = NULL, family = NULL, level = "species",
  formula = "I ~ . * .", traits = NULL, step = FALSE, ...)
}
\arguments{
\item{...}{other params passed to the glm, stepAIC or randomForest algorithm}

\item{an}{object of the class alienData, see as.alienData function.}

\item{method}{to be applied on the data which can be 'rf' for randomForest or 'glm'.}

\item{glm}{family, one family of the `stats` package (e.g. binomial(), Gamma(), gaussian() etc.)}

\item{set}{the formula passed to the algorithm (see `family`).}

\item{specify}{the ecological level on which the method is applied. Choices are 'individus' or 'species'. Default is species.}

\item{vector}{of traits to include in the algorithm}

\item{logical, }{specify if variables are selected by stepAIC. Only for glm}
}
\description{
Fit direct matching centrality models among species or individus
}
\examples{
fitDirectCentrality(bartomeus,family='poisson',level='species')
}
\author{
Dominique Gravel & Steve Vissault
}
\references{
Rohr
}

