% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDirectCentrality.R
\name{fitDirectCentrality}
\alias{fitDirectCentrality}
\title{Fit Direct matching centrality among species or individus}
\usage{
fitDirectCentrality(data, family = "binomial", level = "species",
  formula = "I ~ . * .", ...)
}
\arguments{
\item{data}{a dataframe with a first column named I (the interaction) and extra columns with covariates (can be named at will)}

\item{...}{other params passed to the glm or randomForest algorithm}

\item{specify}{the algorithm: one of the `stats` family or the classification algorithm `randomForest`. Default is binomial.}

\item{specify}{the ecological level on which the method is applied. Choices are 'individus' or 'species'. Default is species.}

\item{set}{the formula passed to the algorithm (see `family`)}
}
\description{
Fit direct matching centrality models among species or individus
}
\examples{
fitDirectCentrality(bartomeus,family='poisson',level='species')
}
\author{
Dominique Gravel & Steve Vissault
}
\references{
Poisot T., Cirtwill A. R., Cazelles A. R., Cazelles K., Gravel D., Fortin M.-J. and Stouffer D.B. (2015). The structure of probabilistic networks. \emph{bioRxiv 016485}; doi: https://doi.org/10.1101/016485.
}

