% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDMC.R
\name{fitDMC}
\alias{fitDMC}
\title{Fit direct matching centrality among species or individus}
\usage{
fitDMC(data, class = NULL, family = NULL, formula = "I ~ . * .",
  level = "species", traits = NULL, step = FALSE, ...)
}
\arguments{
\item{data}{an object of the class alienData, see as.alienData function.}

\item{class}{method to be applied on the data which can be 'rf' for randomForest or 'glm'.}

\item{family}{one family of the \code{stats} package (e.g. binomial(), Gamma(), gaussian() etc.)}

\item{formula}{set the formula passed to the algorithm (see \code{family}).}

\item{level}{specify the ecological level on which the method is applied. Choices are 'individus' or 'species'. Default is species.}

\item{traits}{vector of traits to include in the algorithm}

\item{step}{logical, specify if variables are selected by stepAIC. Only availabe for glm class. FALSE by default}

\item{...}{other params passed to the glm, step or randomForest functions.}
}
\description{
Fit direct matching centrality models among species or individus
}
\author{
Dominique Gravel & Steve Vissault
}
