% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitIMC.R
\name{fitIMC}
\alias{fitIMC}
\title{Fit indirect matching centrality model}
\usage{
fitIMC(data, d = 2, verbose = TRUE, control = list())
}
\arguments{
\item{data}{An object of the class \code{\link{alienData}}.}

\item{d}{Numeric. The dimension of the latent traits. Default is 2.}

\item{verbose}{Logical. Whether information on the progress of the analysis is reported in the console.}

\item{control}{List passed to \code{\link[GenSA]{GenSA}} to control the behavior of the algorithm.}
}
\value{
An object with a class alienFit and a class fitIMC.
}
\description{
This method estimate matching and centrality latent traits to model the interactions in an adjacency matrix.
}
\details{
As can be hinted by the name of the method, there are two types of latent traits.  : (1) matching latent traits that are designed to quantify the strength of the interaction between two species and (2) centrality latent traits, which quantify the number of relations a species has with other species. Mathematically, these latent traits (both matching and centrality) are all orthonormal with each other, within and outside of their category.

When deciding on the dimension of the lantent traits, aside from technical issues (i.e. the number of parameters to estimate and the size of the data) it is important to also consider what is gained (or loss) from increasing (or decreasing) the dimension of the latent traits. The default was set to 2 because it is often of interest to study latent traits in pairs in an ordination-type graphic.
}
\references{
Rohr, R. P. & Bascompte, J. (2014) Components of Phylogenetic Signal in Antagonistic and Mutualistic Networks. Am. Nat. 184, 556--564.

Rohr, R. P., Naisbit, R. E., Mazza, C. & Bersier, L.-F. (2016) Matching-centrality decomposition and the forecasting of new links in networks. Proc. R. Soc. B Biol. Sci. 283, 20152702.
}
\author{
Kevin Cazelles, Dominique Gravel and F. Guillaume Blanchet
}
