% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitKNN.R
\name{fitKNN}
\alias{fitKNN}
\title{Fit using K-nearest neighbour}
\usage{
fitKNN(data, distFrom = "jaccard", distTo = "jaccard",
  distTraitFrom = "euclidean", distTraitTo = "euclidean", traitWeight,
  nNeig)
}
\arguments{
\item{data}{An object of the class alienData, see \code{\link{alienData}}.}

\item{distFrom}{Character string defining which distance (or dissimilarity) to apply on the "From" species. Check \code{\link[vegan]{vegdist}} for the distances to choose from. Default is "jaccard".}

\item{distTo}{Character string defining which distance (or dissimilarity) to apply on the "To" species. Check \code{\link[vegan]{vegdist}} for the distances to choose from. Default is "jaccard".}

\item{distTraitFrom}{Character string defining which distance (or dissimilarity) to apply on traits of the "From" species. Check \code{\link[vegan]{vegdist}} for the distances to choose from. Default is "euclidean".}

\item{distTraitTo}{Character string defining which distance (or dissimilarity) to apply on traits of the "To" species. Check \code{\link[vegan]{vegdist}} for the distances to choose from. Default is "euclidean".}

\item{traitWeight}{Numeric. Defines the contribution of the traits in the analysis. Must be between 0 and 1. Default is 0.5.}

\item{nNeig}{Integer defining how many neighbours to consider.}
}
\value{
A matrix giving the the probability of interaction among each pairs of species.
}
\description{
Model adegency matrix using K-nearest neighbour approach
}
\details{
This function should only be used for bipartite adjency matrice.

The argument \code{traitWeight} defines the important of traits in the analysis. If a weight of 0 is given, the traits are assumes to have no importance. Conversely, if \code{traitWeight} is 1 the traits are given their full importance in the analysis.

When ranking the species to find the \code{nNeig} nearest neighbour, in case of ties the argument   \code{tie.method} in the \code{\link{rank}} function is set to "first".
}
\author{
F. Guillaume Blanchet and Dominique Gravel
}
