% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMC.R
\name{fitMC}
\alias{fitMC}
\title{Esimation of iteractions probability using the matching centrality
approach.}
\usage{
fitMC(data, d = 1, mxt = 10, verbose = TRUE)
}
\arguments{
\item{data}{an object of the class alienData or an adjacency matrix.}

\item{d}{dimensionnality.}

\item{mxt}{Numeric. Maximum running time in seconds.}

\item{verbose}{Logical. Should extra information be reported on progress?}
}
\value{
An object of class \code{alienPredict}.
}
\description{
Esimation of iteractions probability using the matching centrality
approach as described in Rohr 2014 and Rohr 2016.
}
\details{
\code{fitMC} implements the matching-centrality method as described in
Rohr 2014 and Rohr 2016. Briefly, the method uses latent traits to fit
interactions. The first step requires the definition of two sets of interacting
species: set1 and set2 (respectively of size \code{nset1} and \code{nset2}).
The first category of latent trait are the matching traits. (to be developped)
The second category include the centrality terms. (to be developped).
Latents traits are buitl under certain topological constraints:
1- all vectors belongs to the orthogonal basis of the unit vector.
2- all matching vector of a particluar set of specues are orthogonal to each other.
}
\examples{
load_all()
set.seed(1987)
n1 = 12
n2 = 18
trait1 <- runif(n1)
trait2 <- runif(n2)
mat <- matrix(0, n1, n2)
for (i in 1:n1) {
  for (j in 1:n2) {
    mat[i, j] <- (trait1[i]-trait2[j])^2
  }
  if (sum(mat[i,]==0)) mat[i,1+floor(runif(1,0,n2))] <- 1
}

res = fitMC(mat)

plot(trait1, res$methodsSpecific$params$M1)
plot(trait2, res$methodsSpecific$params$M2)

}
\references{
Rohr, R. P. & Bascompte, J. Components of Phylogenetic Signal in Antagonistic and Mutualistic Networks. Am. Nat. 184, 556--564 (2014).

Rohr, R. P., Naisbit, R. E., Mazza, C. & Bersier, L.-F. Matching-centrality decomposition and the forecasting of new links in networks. Proc. R. Soc. B Biol. Sci. 283, 20152702 (2016).
}
\author{
Kevin Cazelles
}
