% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMC.R
\name{fitMC}
\alias{fitMC}
\title{Matching centrality as described in Rohr 2014 and 2016.}
\usage{
fitMC(data, d = 1, mxt = 10)
}
\arguments{
\item{data}{an object of the class alienData, see \code{alienData} function.}

\item{d}{dimensionnality.}

\item{mxt}{Numeric. Maximum running time in seconds.}
}
\description{
Fit matching centrality models among species or individus
}
\references{
Rohr, R. P. & Bascompte, J. Components of Phylogenetic Signal in Antagonistic and Mutualistic Networks. Am. Nat. 184, 556--564 (2014).

Rohr, R. P., Naisbit, R. E., Mazza, C. & Bersier, L.-F. Matching-centrality decomposition and the forecasting of new links in networks. Proc. R. Soc. B Biol. Sci. 283, 20152702 (2016).
}
\author{
Kevin Cazelles
}
