% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitPNB.R
\name{fitPNB}
\alias{fitPNB}
\title{Fit using Probabilistic niche model}
\usage{
fitPNB(
  data,
  type,
  optimum,
  optimumMin,
  optimumMax,
  range,
  rangeMin,
  rangeMax,
  na.rm = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{data}{An object of the class alienData, see \code{\link{alienData}}.}

\item{type}{Method to be used to estimate the model. Either 'P' (presence-only) or 'PA' (presence-absence), respectively.}

\item{optimum}{A vector of two values defining the optimum intercept (first value) and slope (second value).}

\item{optimumMin}{A vector of two values giving the minimum values the optimum intercept and slope can have, respectively.}

\item{optimumMax}{A vector of two values giving the maximum values the optimum intercept and slope can have, respectively.}

\item{range}{A vector of two values defining the range intercept (first value) and slope (second value).}

\item{rangeMin}{A vector of two values giving the minimum values the range intercept and slope can have, respectively.}

\item{rangeMax}{A vector of two values giving the maximum values the range intercept and slope can have, respectively.}

\item{verbose}{Logical. Whether messages from the algorithm are shown (TRUE) or not (FALSE). Default is TRUE.}
}
\value{
An object of class alienFit.
}
\description{
Model adegency matrix using probabilistic niche model
}
\details{
This function is only designed to handle presence-only and presence-absence data. In addition, the function can only handle a single continuous trait for each species.

If there are any NAs in the species interaction data (the adjacency matrix), they will be automatically removed to estimate the presence-absence (PA) model parameters. If there are NAs in the traits, an error message will be sent.
}
\author{
Dominique Gravel and F. Guillaume Blanchet
}
