% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdjacencyMatrix.R
\name{getAdjacencyMatrix}
\alias{getAdjacencyMatrix}
\title{Compute the adjacency matrix for a given alienData object}
\usage{
getAdjacencyMatrix(object, bipartite = FALSE, binary = FALSE)
}
\arguments{
\item{object}{An object of class \code{alienData}.}

\item{bipartite}{A logical. Is the adjacency matrix associated with a
bipartite network? Default is set to \code{FALSE}.}

\item{binary}{A logical. Should the interactions be binary (see details)?
Otherwise the \code{value} column of \code{dfEdges} is used. Default is set
to \code{FALSE}.}
}
\value{
An adjacency matrix of class \code{matrix}.
}
\description{
\code{getAdjacencyMatrix} compute the adjacency matrix based on
the \code{dfEdges} element of an alienData object.
}
\details{
By default \code{getAdjacencyMatrix} creates a square matrix including all
nodes found in the \code{dfNodes} data frame of \code{object}. Then, it reads
\code{dfEdges} to fill the matrix out. If bipartite is \code{TRUE} then only the
nodes found in the \code{idFrom} column of \code{dfEdges} are used to name
the rows of the adjacency matrix. Similarly, \code{idTo}'s nodes become the
columns' names.

Currenlty if there are several values for the same interaction
\code{getAdjacencyMatrix} sums them unless \code{binary} is \code{TRUE}.
}
\author{
Kevin Cazelles
}
\keyword{adjacency}
\keyword{matrix}

