% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getNull}
\alias{getNull}
\alias{getNullOne}
\alias{prodNorm}
\title{getNull}
\usage{
getNull(A)

getNullOne(nbsp)

prodNorm(nbsp, B, V)
}
\arguments{
\item{A}{a matrix.}

\item{nbsp}{an integer (from an ecological standpoint, a number
of species).}

\item{B}{a matrix (it should actually be a orthogonal basis).}

\item{V}{a vector.}
}
\value{
The null basis.
}
\description{
A wrapper around the \code{null} function of Armadillo library
that finds the orthonormal basis of the null space of matrix A. Note that
this wrapper do not allow the user to use any other parameters the
original function can.
}
\section{Functions}{
\itemize{
\item \code{getNullOne}: Special case of \code{getNull} for which the orthonormal
basis of the unit vector is required.

\item \code{prodNorm}: A repeated matrix produc in \code{fitMC} that requires
to normalise one vector.
}}

\author{
Kevin Cazelles
}
