% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTraitMatrix.R
\name{getTraitMatrix}
\alias{getTraitMatrix}
\title{Compute the trait matrix (or matrices) for a given alienData object}
\usage{
getTraitMatrix(object, bipartite = TRUE)
}
\arguments{
\item{object}{An object of class \code{alienData}.}

\item{bipartite}{Logical. Whether the network to consider is bipartite (TRUE) or not (FALSE). Default is FALSE.}
}
\value{
An object of class alienTrait
}
\description{
Computes the trait matrix (or matrices) using \code{node} and \code{trait} of an \link{alienData} object.
}
\details{
When the argument \code{bipartite} is set to TRUE, it is assumed that there are two sets of species with potentially different sets of traits. In this case, \code{edge} will be used in the \code{alienData} object to seperate both sets of traits
}
\author{
F. Guillaume Blanchet
}
