% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{medfish}
\alias{medfish}
\title{Discoveries Of Native And Alien Fish Species In The Eastern Mediterranean Sea}
\format{
A data frame with 60 rows and 4 columns:
\describe{
\item{year}{Year of record}
\item{time}{Time since the first record in the records data}
\item{natives}{Number of newly recorded native species in corresponding year}
\item{aliens}{Number of newly recorded alien species in corresponding year}
}
}
\source{
\url{https://esajournals.onlinelibrary.wiley.com/doi/abs/10.1890/07-1904.1}

\url{https://www.mapress.com/zt/article/view/zootaxa.4956.1.1}
}
\usage{
medfish
}
\description{
The updated data used in Belmaker et al (2009) describing discoveries of native and alien species in the Mediterranean Sea
between the years 1927–2017. The original data is described in full in Golani (2005); see also Golani (2021) for the updated version used here.
}
\references{
Golani, D. 2005. Checklist of the Mediterranean Fishes of Israel. Zootaxa, 947, 1-90. https://mapress.com/zt/article/view/zootaxa.947.1.1

Belmaker, J., Brokovich, E., China, V., Golani, D., and Kiflawi, M. 2009. Estimating the rate of biological introductions: Lessepsian fishes in the Mediterranean. Ecology, 90(4), 1134–1141. https://esajournals.onlinelibrary.wiley.com/doi/10.1890/07-1904.1

Golani, D. 2021. An updated Checklist of the Mediterranean fishes of Israel, with illustrations of recently recorded species and delineation of Lessepsian migrants. Zootaxa, 4956, 1-108. https://www.mapress.com/zt/article/view/zootaxa.4956.1.1
}
\keyword{datasets}
