% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.bayesreg.R
\name{predict.bayesreg}
\alias{bayesian}
\alias{predict.bayesreg}
\alias{prediction}
\alias{regression}
\title{Predict species interaction probabilities based on a fitted hierarchical bayesian regression}
\usage{
\method{predict}{bayesreg}(x, newdata = NULL)
}
\arguments{
\item{x}{A JAGS model file return from fit.bayesreg}

\item{newdata}{A data frame with columns "I","J" and "Traitmatch"}
}
\value{
A vector of probability of interaction values.
I: Species in the higher level (eg. pollinators)
J: Species in the lower level (eg. plants)
mean: Mean probability of interaction
lower: 0.05 quantile of the probability of interaction
upper: 0.95 quantile of the probability of interaction
}
\description{
A helper functional for generating the probability of interaction among hypothetical species based on their traits.
The trait-matching function is currently |Trait_i - Trait_j|
}
\details{
The newdata object gives the trait-matching between species I (eg. pollinator) and species J (eg. plant)
If fit.bayesreg was fit to an intercept model, newdata is ignored, not possible to create a prediction of a new species. In this case, it is assumed that the user wants the probability matrix for all input species.
}
\author{
Ben Weinstein
}
\references{
Weinstein et al. 2016. Oikos
}

